requirements.txt
streamlit
audiorecorder
whisper
pyttsx3
pydub
google-generativeai
python-dotenv
torch
torchvision
torchaudio
